split reader settings into multiple components
