+++ src/renderer/reader/components/ReaderSettings/DivinaSetReadingMode.tsx
@@ -1,0 +1,108 @@
+// ==LICENSE-BEGIN==
+// Copyright 2017 European Digital Reading Lab. All rights reserved.
+// Licensed to the Readium Foundation under one or more contributor license agreements.
+// Use of this source code is governed by a BSD-style license
+// that can be found in the LICENSE file exposed on Github (readium) in the project repository.
+// ==LICENSE-END==
+
+import * as stylesReader from "readium-desktop/renderer/assets/styles/reader-app.scss";
+
+import * as React from "react";
+
+import SVG from "readium-desktop/renderer/common/components/SVG";
+import * as DoneIcon from "readium-desktop/renderer/assets/icons/done.svg";
+import { useTranslator } from "readium-desktop/renderer/common/hooks/useTranslator";
+import { IReaderSettingsProps } from "readium-desktop/renderer/reader/components/options-values";
+
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+interface IBaseProps extends IReaderSettingsProps {
+    // handleSettingsClick: (open: boolean) => void;
+
+    // tabValue: string;
+    // setTabValue: (value: string) => void;
+}
+
+export const DivinaSetReadingMode = ({ handleDivinaReadingMode, divinaReadingMode, divinaReadingModeSupported }: Partial<IBaseProps>) => {
+    const [__] = useTranslator();
+
+    return (
+        <div id={stylesReader.themes_list} aria-label={__("reader.settings.disposition.title")} role="radiogroup">
+            <div>
+                <input
+                    disabled={!divinaReadingModeSupported.includes("double")}
+                    id={"radio-" + "double"}
+                    type="radio"
+                    name="divinaReadingMode"
+                    onChange={() => {
+                        handleDivinaReadingMode("double");
+                    }}
+                    checked={divinaReadingMode === "double"}
+                />
+                <label
+                    aria-disabled={!divinaReadingModeSupported.includes("double")}
+                    htmlFor={"radio-" + "double"}
+                >
+                    {divinaReadingMode === "double" && <SVG svg={DoneIcon} ariaHidden />}
+                    {"double"}
+                </label>
+            </div>
+            <div>
+                <input
+                    disabled={!divinaReadingModeSupported.includes("guided")}
+                    id={"radio-" + "guided"}
+                    type="radio"
+                    name="divinaReadingMode"
+                    onChange={() => {
+                        handleDivinaReadingMode("guided");
+                    }}
+                    checked={divinaReadingMode === "guided"}
+                />
+                <label
+                    aria-disabled={!divinaReadingModeSupported.includes("guided")}
+                    htmlFor={"radio-" + "guided"}
+                >
+                    {divinaReadingMode === "guided" && <SVG svg={DoneIcon} ariaHidden />}
+                    {"guided"}
+                </label>
+            </div>
+            <div>
+                <input
+                    disabled={!divinaReadingModeSupported.includes("scroll")}
+                    id={"radio-" + "scroll"}
+                    type="radio"
+                    name="divinaReadingMode"
+                    onChange={() => {
+                        handleDivinaReadingMode("scroll");
+                    }}
+                    checked={divinaReadingMode === "scroll"}
+                />
+                <label
+                    aria-disabled={!divinaReadingModeSupported.includes("scroll")}
+                    htmlFor={"radio-" + "scroll"}
+                >
+                    {divinaReadingMode === "scroll" && <SVG svg={DoneIcon} ariaHidden />}
+                    {"scroll"}
+                </label>
+            </div>
+            <div>
+                <input
+                    disabled={!divinaReadingModeSupported.includes("single")}
+                    id={"radio-" + "single"}
+                    type="radio"
+                    name="divinaReadingMode"
+                    onChange={() => {
+                        handleDivinaReadingMode("single");
+                    }}
+                    checked={divinaReadingMode === "single"}
+                />
+                <label
+                    aria-disabled={!divinaReadingModeSupported.includes("single")}
+                    htmlFor={"radio-" + "single"}
+                >
+                    {divinaReadingMode === "single" && <SVG svg={DoneIcon} ariaHidden />}
+                    {"single"}
+                </label>
+            </div>
+        </div>
+    );
+};
