+++ src/renderer/reader/components/ReaderSettings/ReadingSpacing.tsx
@@ -1,0 +1,92 @@
+// ==LICENSE-BEGIN==
+// Copyright 2017 European Digital Reading Lab. All rights reserved.
+// Licensed to the Readium Foundation under one or more contributor license agreements.
+// Use of this source code is governed by a BSD-style license
+// that can be found in the LICENSE file exposed on Github (readium) in the project repository.
+// ==LICENSE-END==
+
+import * as stylesSettings from "readium-desktop/renderer/assets/styles/components/settings.scss";
+
+import * as React from "react";
+import { useTranslator } from "readium-desktop/renderer/common/hooks/useTranslator";
+import { useSelector } from "readium-desktop/renderer/common/hooks/useSelector";
+import { IReaderRootState } from "readium-desktop/common/redux/states/renderer/readerRootState";
+import { readerConfigInitialState } from "readium-desktop/common/redux/states/reader";
+import { useSavePublisherReaderConfigDebounced } from "readium-desktop/renderer/common/hooks/useReaderConfig";
+import { Slider } from "readium-desktop/renderer/reader/components/ReaderSettings/Slider";
+import { AdjustableSettingsStrings } from "readium-desktop/renderer/reader/components/options-values";
+
+interface ITable {
+    title: string,
+    ariaLabel: string,
+    ariaValuemin: number,
+    defaultValue: string,
+    parameter: keyof AdjustableSettingsStrings;
+    altParameter: string,
+    rem: boolean,
+}
+
+export const ReadingSpacing = () => {
+
+    const [__] = useTranslator();
+
+    const set = useSavePublisherReaderConfigDebounced();
+    const config = useSelector((state: IReaderRootState) => state.reader.config);
+    const { pageMargins, wordSpacing, letterSpacing, paraSpacing, lineHeight } = config;
+    const spacingOptions: ITable[] = [
+        {
+            title: `${__("reader.settings.margin")}`,
+            ariaLabel: "label_pageMargins",
+            ariaValuemin: 0,
+            defaultValue: pageMargins,
+            parameter: "pageMargins",
+            altParameter: `${readerConfigInitialState.pageMargins}`,
+            rem: false,
+        },
+        {
+            title: `${__("reader.settings.wordSpacing")}`,
+            ariaLabel: "label_wordSpacing",
+            ariaValuemin: 0,
+            defaultValue: wordSpacing,
+            parameter: "wordSpacing",
+            altParameter: `${readerConfigInitialState.wordSpacing}`,
+            rem: true,
+        },
+        {
+            title: `${__("reader.settings.letterSpacing")}`,
+            ariaLabel: "label_letterSpacing",
+            ariaValuemin: 0,
+            defaultValue: letterSpacing,
+            parameter: "letterSpacing",
+            altParameter: `${readerConfigInitialState.letterSpacing}`,
+            rem: true,
+        },
+        {
+            title: `${__("reader.settings.paraSpacing")}`,
+            ariaLabel: "label_paraSpacing",
+            ariaValuemin: 0,
+            defaultValue: paraSpacing,
+            parameter: "paraSpacing",
+            altParameter: `${readerConfigInitialState.paraSpacing}`,
+            rem: true,
+        },
+        {
+            title: `${__("reader.settings.lineSpacing")}`,
+            ariaLabel: "label_lineHeight",
+            ariaValuemin: 0,
+            defaultValue: lineHeight,
+            parameter: "lineHeight",
+            altParameter: `${readerConfigInitialState.lineHeight}`,
+            rem: true,
+        },
+    ];
+
+    return (
+        <div className={stylesSettings.settings_tab_container_reading_spacing}>
+            {spacingOptions.map((option: ITable) => (
+                <Slider value={config[option.parameter]} option={option} key={option.title} set={set} />
+            ))}
+        </div>
+    );
+
+};
